+++ CHANGELOG.md
@@ -11,8 +11,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.1.1] - 2025-05-05
 - fix #2, wrong color
 - fix version number in library.properties
-- writeRegister / readRegister full name.
-
+- give **writeRegister() / readRegister()** full name.
+- add **PCA9632_AUTO_INCR_ALL** in **write(array)**
+- add **allOff()**
+- add **getGroupPWM() + setGroupPWM(uint8_t value)**
+- add **getGroupFREQ() + setGroupFREQ(uint8_t value)**
+- update readme.md
+- minor edits
 
 
 ## [0.1.0] - 2024-11-25